+++ docs/pages/components/progress/examples/ExSimple.vue
@@ -1,0 +1,11 @@
+<template>
+    <section>
+        <b-progress />
+
+        <b-progress value="20" />
+
+        <b-progress value="40" show-value />
+
+        <b-progress value="80" show-value format="percent" />
+    </section>
+</template>